commit: refactor mudando nome
