+++ src/main.js
@@ -3,7 +3,7 @@ import e from "express";
 import user_router from "./http/routes/user-route.js";
 import roullete_bet_router from "./http/routes/roullete-bet-route.js";
 import jackpot_bet_router from "./http/routes/jackpot-bet-route.js";
-import prize_bet_router from "./http/routes/prize-bet-router.js";
+import prize_bet_router from "./http/routes/prize-bet-router-kayke.js";
 const app = e();
 
 app.use(e.json());
